test: cover the tail path and keep the large-input test off -race
The regression test added with the previous commit had two problems.

Both of its lengths were exact multiples of the write block, so bufused
was always zero when Sum32 ran and the two tail loops never executed at
a length that does not fit in 32 bits. It pinned exactly the bug that
was fixed and nothing adjacent to it. The second case is now 4 GiB + 5
bytes, with the remainder written first so that every subsequent write
also carries buffered bytes across the call boundary. Against the
unfixed code the two cases now fail with different wrong values,
2cc5d05 and 1295514d, rather than with the same one.

It also cost about 78s under -race, against 1.3s without. There is no
assembly for updateGo on amd64, so the race detector instruments every
four-byte load in its hot loop and the test pushes several GiB through
it. ci.yml runs the suite twice with -race across a three-OS by
three-Go-version matrix, so that added roughly 23 minutes per pull
request to a suite which otherwise finishes in under two seconds. The
test has moved to its own file behind //go:build !race; it exercises no
concurrency, so the race detector has nothing to check in it. The added
cost is now about 1.9s per job.

Finally, running gofmt over xxh32zero_test.go had reformatted the
benchmark banner comment, which is unrelated churn. That file is back to
its original contents.

Testing: go test ./... passes in all four combinations ci.yml runs. With
the fix reverted to the v4.1.28 version, the new test fails both cases;
with the fix in place it passes.
